+++ extensions/social-network-trends/src/utils/common-utils.ts
@@ -1,5 +1,11 @@
-export const isEmpty = (string: string | null | undefined) => {
-  return !(string != null && String(string).length > 0);
+import { getPreferenceValues, LocalStorage } from "@raycast/api";
+import Values = LocalStorage.Values;
+
+export const commonPreferences = () => {
+  const preferencesMap = new Map(Object.entries(getPreferenceValues<Values>()));
+  return {
+    rememberTag: preferencesMap.get("rememberTag") as boolean,
+  };
 };
 
 export const listIcon = [