+++ extensions/social-network-trends/CHANGELOG.md
@@ -1,5 +1,5 @@
 # Social Network Trends Changelog
 
-## [Initial Version] - 2022-04-25
+## [Initial Version] - 2022-04-30
 
-Show popular trends or Search something on WeiBo, ZhiHu, and DouYin.
+- Show trends on WeiBo, ZhiHu, and DouYin.